feat: add all stores
