+++ Features/Sources/ContributionModule/ContributionStore+Extensions.swift
@@ -6,91 +6,11 @@
 //
 
 import Foundation
-import NetworkKit
-import StatsKit
 import CoreModule
-import EventModule
 import Models
 import Stores
-import NetworkModule
-
-public extension ContributionStore {
-    
-    @MainActor
-    func fetchContributions(savingsplanID: Int) async {
-        do {
-            let contributions = try await NetworkService.sendRequest(
-                apiBuilder: ContributionAPIRequester.fetch(savingsplanID: savingsplanID),
-                responseModel: [ContributionModel].self
-            )
-            self.contributions = contributions
-            sortContributionsByDate()
-        } catch { NetworkService.handleError(error: error) }
-    }
-    
-    @discardableResult
-    @MainActor
-    func createContribution(savingsplanID: Int, body: ContributionModel) async -> ContributionModel? {
-        do {
-            let response = try await NetworkService.sendRequest(
-                apiBuilder: ContributionAPIRequester.create(savingsplanID: savingsplanID, body: body),
-                responseModel: ContributionResponseWithAmount.self
-            )
-            if let contribution = response.contribution, let newAmount = response.newAmount {
-                self.contributions.append(contribution)
-                sortContributionsByDate()
-                SavingsPlanStore.shared.setNewAmount(savingsPlanID: savingsplanID, newAmount: newAmount)
-                EventService.sendEvent(key: EventKeys.contributionCreated)
-                return contribution
-            }
-            
-            return nil
-        } catch {
-            NetworkService.handleError(error: error)
-            return nil
-        }
-    }
-    
-    @MainActor
-    func updateContribution(savingsplanID: Int, contributionID: Int, body: ContributionModel) async {
-        do {
-            let response = try await NetworkService.sendRequest(
-                apiBuilder: ContributionAPIRequester.update(savingsplanID: savingsplanID, contributionID: contributionID, body: body),
-                responseModel: ContributionResponseWithAmount.self
-            )
-            if let contribution = response.contribution, let newAmount = response.newAmount {
-                if let index = self.contributions.map(\.id).firstIndex(of: contributionID) {
-                    self.contributions[index] = contribution
-                    sortContributionsByDate()
-                    SavingsPlanStore.shared.setNewAmount(savingsPlanID: savingsplanID, newAmount: newAmount)
-                    EventService.sendEvent(key: EventKeys.contributionUpdated)
-                }
-            }
-        } catch { NetworkService.handleError(error: error) }
-    }
-    
-    @MainActor
-    func deleteContribution(savingsplanID: Int, contributionID: Int) async {
-        do {
-            let response = try await NetworkService.sendRequest(
-                apiBuilder: ContributionAPIRequester.delete(savingsplanID: savingsplanID, contributionID: contributionID),
-                responseModel: ContributionResponseWithAmount.self
-            )
-            if let newAmount = response.newAmount {
-                SavingsPlanStore.shared.setNewAmount(savingsPlanID: savingsplanID, newAmount: newAmount)
-            }
-            self.contributions.removeAll(where: { $0.id == contributionID })
-            EventService.sendEvent(key: EventKeys.contributionDeleted)
-        } catch { NetworkService.handleError(error: error) }
-    }
-    
-}
 
 extension ContributionStore {
-    
-    private func sortContributionsByDate() {
-        self.contributions.sort { $0.date > $1.date }
-    }
     
     public func getContributionsAmountByMonth(for year: Int) -> [Double] {
         var contributionsByMonth = Array(repeating: 0.0, count: 12)
