+++ Features/Sources/TransactionModule/Stores/TransactionStore+Utils.swift
@@ -21,10 +21,7 @@ public extension TransactionStore {
         return self.transactions
             .filter { $0.subcategory?.id == subcategoryID }
     }
-    
-    func sortTransactionsByDate() {
-        self.transactions.sort { $0.date > $1.date }
-    }
+  
 }
 
 extension TransactionStore {
