+++ Features/Sources/TransactionModule/Stores/TransactionStore.swift
@@ -45,106 +45,6 @@ public extension TransactionStore {
 
 public extension TransactionStore {
     
-    @MainActor
-    func fetchTransactionsByPeriod(accountID: Int, startDate: Date, endDate: Date, type: TransactionType? = nil) async {
-        guard dateFetched.filter({ Calendar.current.isDate($0, equalTo: startDate, toGranularity: .month) }).isEmpty else { return }
-        
-        do {
-            let transactions = try await TransactionService.fetchTransactionsByPeriod(
-                accountID: accountID,
-                startDate: startDate,
-                endDate: endDate,
-                type: type
-            ).map { try $0.toModel() }
-            
-            currentDateForFetch = startDate
-            self.dateFetched.append(currentDateForFetch)
-            
-            self.transactions += transactions
-            sortTransactionsByDate()
-            
-            EventService.sendEvent(key: EventKeys.transactionListPagination)
-        } catch { NetworkService.handleError(error: error) }
-    }
-    
-    /// Create transaction, add it to repository and optionally return it
-    @discardableResult
-    @MainActor
-    func createTransaction(accountID: Int, body: TransactionDTO, shouldReturn: Bool = false, addInRepo: Bool = true) async -> TransactionModel? {
-        do {
-            let response = try await TransactionService.create(accountID: accountID, body: body)
-            
-            if let transaction = try response.transaction?.toModel(), let newBalance = response.newBalance {
-                if addInRepo {
-                    self.transactions.append(transaction)
-                    sortTransactionsByDate()
-                }
-                AccountStore.shared.setNewBalance(accountID: accountID, newBalance: newBalance)
-                EventService.sendForTransactionCreated(transaction: transaction)
-                return shouldReturn ? transaction : nil
-            }
-            return nil
-        } catch {
-            NetworkService.handleError(error: error)
-            return nil
-        }
-    }
-    
-    /// Create transaction and optionally return it
-    @discardableResult
-    @MainActor
-    func updateTransaction(accountID: Int, transactionID: Int, body: TransactionDTO, shouldReturn: Bool = false) async -> TransactionModel? {
-        do {
-            let response = try await TransactionService.update(transactionID: transactionID, body: body)
-            
-            if let transaction = try response.transaction?.toModel(), let newBalance = response.newBalance {
-                if let index = self.transactions.map(\.id).firstIndex(of: transaction.id) {
-                    self.transactions[index] = transaction
-                    sortTransactionsByDate()
-                    AccountStore.shared.setNewBalance(accountID: accountID, newBalance: newBalance)
-                    EventService.sendEvent(key: EventKeys.transactionUpdated)
-                    return shouldReturn ? transaction : nil
-                }
-            }
-            return nil
-        } catch {
-            NetworkService.handleError(error: error)
-            return nil
-        }
-    }
-    
-    @MainActor
-    func fetchCategory(name: String, transactionID: Int? = nil) async -> TransactionFetchCategoryResponse? {
-        do {
-            return try await TransactionService.fetchRecommendedCategory(name: name, transactionID: transactionID)
-        } catch {
-            NetworkService.handleError(error: error)
-            return nil
-        }
-    }
-    
-    @MainActor
-    func deleteTransaction(transactionID: Int) async {
-        let accountRepo: AccountStore = .shared
-        do {
-            let response = try await TransactionService.delete(transactionID: transactionID)
-            
-            if let index = self.transactions.firstIndex(where: { $0.id == transactionID }) {
-                self.transactions.remove(at: index)
-            }
-            TransferStore.shared.transfers.removeAll { $0.id == transactionID }
-            
-            if let newBalance = response.newBalance, let account = accountRepo.selectedAccount, let accountID = account._id {
-                AccountStore.shared.setNewBalance(accountID: accountID, newBalance: newBalance)
-            }
-            
-            EventService.sendEvent(key: EventKeys.transactionDeleted)
-        } catch { NetworkService.handleError(error: error) }
-    }
-}
-
-public extension TransactionStore {
-    
     func fetchTransactionsOfCurrentMonth(accountID: Int) async {
         let startDate = Date().startOfMonth ?? .now
         let endDate = Date().endOfMonth ?? .now
@@ -235,43 +135,5 @@ public extension TransactionStore {
     func getTransactionFromSubscriptions(in month: Date? = nil) -> [TransactionModel] {
         return getTransactions(in: month)
             .filter { $0.isFromSubscription == true }
-    }
-}
-
-public extension TransactionDTO {
-    
-    func toModel() throws -> TransactionModel {
-        guard let id,
-              let amount,
-              let dateISO
-        else { throw NetworkError.unknown }
-                
-        let date = dateISO.toDate()
-        let category = CategoryStore.shared.findCategoryById(categoryID)
-        
-        let subcategory = CategoryStore.shared.findSubcategoryById(subcategoryID)
-        
-        let senderAccount = AccountStore.shared.findByID(senderAccountID)
-        let receiverAccount = AccountStore.shared.findByID(receiverAccountID)
-        
-        return .init(
-            id: id,
-            name: name ?? "",
-            amount: amount,
-            date: date ?? .now,
-            creationDate: creationDate?.toDate(),
-            category: category,
-            subcategory: subcategory,
-            note: note,
-            isFromSubscription: isFromSubscription ?? false,
-            isFromApplePay: isFromApplePay ?? false,
-            nameFromApplePay: nameFromApplePay,
-            autoCat: autoCat,
-            senderAccount: senderAccount,
-            receiverAccount: receiverAccount,
-            address: address,
-            lat: lat,
-            long: long
-        )
     }
 }
