+++ Models/Sources/Model/Transaction/TransactionDTO.swift
@@ -6,7 +6,6 @@
 //
 
 import Foundation
-import NetworkKit
 
 public struct TransactionDTO: Codable {
     public var id: Int?
